backtest_system/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
psycopg2-binary>=2.9.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
matplotlib>=3.7.0
openpyxl>=3.1.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
